+++ primitive/update.py
@@ -22,6 +22,7 @@ from bpy.props import (StringProperty, IntProperty, FloatProperty,
 
 from .adaptive_plane import Adaptive_Plane
 from .box import Box
+from .bolt import Bolt
 from .capsule import Capsule
 from .cylinder import Cylinder, Cone
 #frome .geoSphere import GeoSphere
@@ -54,6 +55,7 @@ from .light import Compass
 def get_class(name):
 	if name == "Adaptive_Plane": return Adaptive_Plane()
 	elif name == "Box": return Box()
+	elif name == "Bolt": return Bolt()
 	elif name == "Capsule": return Capsule()
 	elif name == "Cone": return Cone()
 	elif name == "Cylinder": return Cylinder()
@@ -206,17 +208,23 @@ class PrimitiveData(PropertyGroup):
 
 	MAX_INPUT_NUMBER = 50
 
-	Bolt : BoolProperty(name = "Bolt", default = True, description = "Bolt")
-	change : BoolProperty(name = "Change",
+	Bolt : BoolProperty(name="Bolt", update = primitive_update,
+			default=True, description="Bolt"
+	)
+	
+	change : BoolProperty(name = "Change", update = primitive_update,
 				default = False,
-				description = "change Bolt")
+				description = "change Bolt"
+	)
 
 	# Model Types
 	Model_Type_List = [('bf_Model_Bolt', 'BOLT', 'Bolt Model'),
-						('bf_Model_Nut', 'NUT', 'Nut Model')]
+						('bf_Model_Nut', 'NUT', 'Nut Model')
+					]
 	bf_Model_Type: EnumProperty(
 			attr='bf_Model_Type',
 			name='Model',
+			update = primitive_update,
 			description='Choose the type off model you would like',
 			items=Model_Type_List, default='bf_Model_Bolt'
 			)
@@ -226,10 +234,12 @@ class PrimitiveData(PropertyGroup):
 						('bf_Head_Cap', 'CAP', 'Cap Head'),
 						('bf_Head_Dome', 'DOME', 'Dome Head'),
 						('bf_Head_Pan', 'PAN', 'Pan Head'),
-						('bf_Head_CounterSink', 'COUNTER SINK', 'Counter Sink Head')]
+						('bf_Head_CounterSink', 'COUNTER SINK', 'Counter Sink Head')
+					]
 	bf_Head_Type: EnumProperty(
 			attr='bf_Head_Type',
 			name='Head',
+			update = primitive_update,
 			description='Choose the type off Head you would like',
 			items=Model_Type_List, default='bf_Head_Hex'
 			)
@@ -237,20 +247,24 @@ class PrimitiveData(PropertyGroup):
 	Bit_Type_List = [('bf_Bit_None', 'NONE', 'No Bit Type'),
 					('bf_Bit_Allen', 'ALLEN', 'Allen Bit Type'),
 					('bf_Bit_Torx', 'TORX', 'Torx Bit Type'),
-					('bf_Bit_Philips', 'PHILLIPS', 'Phillips Bit Type')]
+					('bf_Bit_Philips', 'PHILLIPS', 'Phillips Bit Type')
+				]
 	bf_Bit_Type: EnumProperty(
 			attr='bf_Bit_Type',
 			name='Bit Type',
+			update = primitive_update,
 			description='Choose the type of bit to you would like',
 			items=Bit_Type_List, default='bf_Bit_None'
 			)
 	# Nut Types
 	Nut_Type_List = [('bf_Nut_Hex', 'HEX', 'Hex Nut'),
 					('bf_Nut_Lock', 'LOCK', 'Lock Nut'),
-					('bf_Nut_12Pnt', '12 POINT', '12 Point Nut')]
+					('bf_Nut_12Pnt', '12 POINT', '12 Point Nut')
+				]
 	bf_Nut_Type: EnumProperty(
 			attr='bf_Nut_Type',
 			name='Nut Type',
+			update = primitive_update,
 			description='Choose the type of nut you would like',
 			items=Nut_Type_List, default='bf_Nut_Hex'
 			)
@@ -258,6 +272,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Shank_Length: FloatProperty(
 			attr='bf_Shank_Length',
 			name='Shank Length', default=0,
+			update = primitive_update,
 			min=0, soft_min=0, max=MAX_INPUT_NUMBER,
 			description='Length of the unthreaded shank',
 			unit='LENGTH',
@@ -267,6 +282,7 @@ class PrimitiveData(PropertyGroup):
 			name='Shank Dia', default=3,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
+			update = primitive_update,
 			description='Diameter of the shank',
 			unit='LENGTH',
 			)
@@ -275,6 +291,7 @@ class PrimitiveData(PropertyGroup):
 			name='Bit Depth', default=1.1431535482406616,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
+			update = primitive_update,
 			description='Depth of the Phillips Bit',
 			unit='LENGTH',
 			)
@@ -283,6 +300,7 @@ class PrimitiveData(PropertyGroup):
 			name='Bit Depth', default=1.5,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
+			update = primitive_update,
 			description='Depth of the Allen Bit',
 			unit='LENGTH',
 			)
@@ -291,6 +309,7 @@ class PrimitiveData(PropertyGroup):
 			name='Flat Dist', default=2.5,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
+			update = primitive_update,
 			description='Flat Distance of the Allen Bit',
 			unit='LENGTH',
 			)
@@ -307,12 +326,14 @@ class PrimitiveData(PropertyGroup):
 	bf_Torx_Size_Type: EnumProperty(
 			attr='bf_Torx_Size_Type',
 			name='Torx Size',
+			update = primitive_update,
 			description='Size of the Torx Bit',
 			items=Torx_Size_Type_List, default='bf_Torx_T20'
 			)
 	bf_Torx_Bit_Depth: FloatProperty(
 			attr='bf_Torx_Bit_Depth',
 			name='Bit Depth', default=1.5,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Depth of the Torx Bit',
@@ -321,6 +342,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Hex_Head_Height: FloatProperty(
 			attr='bf_Hex_Head_Height',
 			name='Head Height', default=2,
+			update = primitive_update,
 			min=0, soft_min=0, max=MAX_INPUT_NUMBER,
 			description='Height of the Hex Head',
 			unit='LENGTH',
@@ -328,6 +350,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Hex_Head_Flat_Distance: FloatProperty(
 			attr='bf_Hex_Head_Flat_Distance',
 			name='Flat Dist', default=5.5,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Flat Distance of the Hex Head',
@@ -336,6 +359,7 @@ class PrimitiveData(PropertyGroup):
 	bf_12_Point_Head_Height: FloatProperty(
 			attr='bf_12_Point_Head_Height',
 			name='Head Height', default=3.0,
+			update = primitive_update,
 			min=0, soft_min=0, max=MAX_INPUT_NUMBER,
 			description='Height of the 12 Point Head',
 			unit='LENGTH',
@@ -343,6 +367,7 @@ class PrimitiveData(PropertyGroup):
 	bf_12_Point_Head_Flat_Distance: FloatProperty(
 			attr='bf_12_Point_Head_Flat_Distance',
 			name='Flat Dist', default=3.0,
+			update = primitive_update,
 			min=0.001, soft_min=0,    #limit to 0.001 to avoid calculation error
 			max=MAX_INPUT_NUMBER,
 			description='Flat Distance of the 12 Point Head',
@@ -351,6 +376,7 @@ class PrimitiveData(PropertyGroup):
 	bf_12_Point_Head_Flange_Dia: FloatProperty(
 			attr='bf_12_Point_Head_Flange_Dia',
 			name='12 Point Head Flange Dia', default=5.5,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Flange diameter of the 12 point Head',
@@ -359,6 +385,7 @@ class PrimitiveData(PropertyGroup):
 	bf_CounterSink_Head_Dia: FloatProperty(
 			attr='bf_CounterSink_Head_Dia',
 			name='Head Dia', default=6.300000190734863,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Diameter of the Counter Sink Head',
@@ -367,6 +394,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Cap_Head_Height: FloatProperty(
 			attr='bf_Cap_Head_Height',
 			name='Head Height', default=3,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Height of the Cap Head',
@@ -375,6 +403,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Cap_Head_Dia: FloatProperty(
 			attr='bf_Cap_Head_Dia',
 			name='Head Dia', default=5.5,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Diameter of the Cap Head',
@@ -383,6 +412,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Dome_Head_Dia: FloatProperty(
 			attr='bf_Dome_Head_Dia',
 			name='Dome Head Dia', default=5.599999904632568,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Length of the unthreaded shank',
@@ -391,6 +421,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Pan_Head_Dia: FloatProperty(
 			attr='bf_Pan_Head_Dia',
 			name='Pan Head Dia', default=5.599999904632568,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Diameter of the Pan Head',
@@ -400,6 +431,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Philips_Bit_Dia: FloatProperty(
 			attr='bf_Philips_Bit_Dia',
 			name='Bit Dia', default=1.8199999332427979,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Diameter of the Philips Bit',
@@ -409,6 +441,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Thread_Length: FloatProperty(
 			attr='bf_Thread_Length',
 			name='Thread Length', default=6,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Length of the Thread',
@@ -418,6 +451,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Major_Dia: FloatProperty(
 			attr='bf_Major_Dia',
 			name='Major Dia', default=3,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Outside diameter of the Thread',
@@ -427,6 +461,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Pitch: FloatProperty(
 			attr='bf_Pitch',
 			name='Pitch', default=0.3499999940395355,
+			update = primitive_update,
 			min=0.1, soft_min=0.1,
 			max=7.0,
 			description='Pitch if the thread',
@@ -435,6 +470,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Minor_Dia: FloatProperty(
 			attr='bf_Minor_Dia',
 			name='Minor Dia', default=2.6211137771606445,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Inside diameter of the Thread',
@@ -443,6 +479,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Crest_Percent: IntProperty(
 			attr='bf_Crest_Percent',
 			name='Crest Percent', default=10,
+			update = primitive_update,
 			min=1, soft_min=1,
 			max=90,
 			description='Percent of the pitch that makes up the Crest',
@@ -450,6 +487,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Root_Percent: IntProperty(
 			attr='bf_Root_Percent',
 			name='Root Percent', default=10,
+			update = primitive_update,
 			min=1, soft_min=1,
 			max=90,
 			description='Percent of the pitch that makes up the Root',
@@ -457,6 +495,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Div_Count: IntProperty(
 			attr='bf_Div_Count',
 			name='Div count', default=36,
+			update = primitive_update,
 			min=4, soft_min=4,
 			max=4096,
 			description='Div count determine circle resolution',
@@ -464,6 +503,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Hex_Nut_Height: FloatProperty(
 			attr='bf_Hex_Nut_Height',
 			name='Hex Nut Height', default=2.4000000953674316,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Height of the Hex Nut',
@@ -472,6 +512,7 @@ class PrimitiveData(PropertyGroup):
 	bf_Hex_Nut_Flat_Distance: FloatProperty(
 			attr='bf_Hex_Nut_Flat_Distance',
 			name='Hex Nut Flat Dist', default=5.5,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Flat distance of the Hex Nut',
@@ -480,6 +521,7 @@ class PrimitiveData(PropertyGroup):
 	bf_12_Point_Nut_Height: FloatProperty(
 			attr='bf_12_Point_Nut_Height',
 			name='12 Point Nut Height', default=2.4000000953674316,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Height of the 12 Point Nut',
@@ -489,6 +531,7 @@ class PrimitiveData(PropertyGroup):
 	bf_12_Point_Nut_Flat_Distance: FloatProperty(
 			attr='bf_12_Point_Nut_Flat_Distance',
 			name='12 Point Nut Flat Dist', default=3.0,
+			update = primitive_update,
 			min=0.001, soft_min=0,    #limit to 0.001 to avoid calculation error
 			max=MAX_INPUT_NUMBER,
 			description='Flat distance of the 12 point Nut',
@@ -497,6 +540,7 @@ class PrimitiveData(PropertyGroup):
 	bf_12_Point_Nut_Flange_Dia: FloatProperty(
 			attr='bf_12_Point_Nut_Flange_Dia',
 			name='12 Point Nut Flange Dia', default=5.5,
+			update = primitive_update,
 			min=0, soft_min=0,
 			max=MAX_INPUT_NUMBER,
 			description='Flange diameter of the 12 point Nut',